Converted all floats to doubles to accommodate larger numbers

+++ Assets/Scripts/Building.cs
@@ -31,9 +31,9 @@ public class Building : MonoBehaviour
         return data;
     }
 
-    public float GetIridiumPerTick()
+    public double GetIridiumPerTick()
     {
-        float x = 0;
+        double x = 0;
         foreach (Troop troop in ownedTroops)
         {
             x += troop.GetIridiumPerTick() * troop.structureOwned;
