+++ Assets/Scripts/Troop.cs
@@ -4,14 +4,14 @@ public class Troop
     public string structureName = "Name";
     public int structureLevel = 1;
     public int structureOwned = 0;
-    public float structureBaseCost = 0;
-    public float structureCurrentCost = 0;
-    public float structureBaseIridiumPerSecond = 0;
-    public float structureIridiumMultiplier = 1; //TODO: Make this dependant on the level of the building
-    public float structureCostMultiplier = 1.25f;
-    public float structureCostMultiplierMultiplier = 1;
+    public double structureBaseCost = 0;
+    public double structureCurrentCost = 0;
+    public double structureBaseIridiumPerSecond = 0;
+    public double structureIridiumMultiplier = 1; //TODO: Make this dependant on the level of the building
+    public double structureCostMultiplier = 1.25f;
+    public double structureCostMultiplierMultiplier = 1;
 
-    public Troop(string name, int baseCost, int baseIridiumPerSecond, float iridiumMultiplier, float costMultiplier, float costMultiplierMultiplier)
+    public Troop(string name, int baseCost, int baseIridiumPerSecond, double iridiumMultiplier, double costMultiplier, double costMultiplierMultiplier)
     {
         structureName = name;
         structureOwned = 0;
@@ -33,9 +33,9 @@ public class Troop
         structureCostMultiplierMultiplier = so.structureCostMultiplierMultiplier;
     }
 
-    public float GetIridiumPerTick()
+    public double GetIridiumPerTick()
     {
-        float x = structureBaseIridiumPerSecond * structureIridiumMultiplier * (1.0f / GameManager.ticksPerSecond);
+        double x = structureBaseIridiumPerSecond * structureIridiumMultiplier * (1.0f / GameManager.ticksPerSecond);
         return x;
     }
 }
